🚚(frontend) rename base backend in dummy backend
On frontend, the base backend aims to be used for development purpose in order
to fake data to prevent to have to launch other services. But the name "base" is
 not explicit about that so we decided to rename it "dummy".
