+++ env.d/development/dev
@@ -1,7 +1,8 @@
 # Authentication Backend
 AUTHENTICATION_BASE_URL=http://localhost:8073
-AUTHENTICATION_BACKEND=base
+AUTHENTICATION_BACKEND=dummy
 
 # LMS Backend
 EDX_BASE_URL=http://localhost:8073
 EDX_BACKEND=richie.apps.courses.lms.base.BaseLMSBackend
+EDX_JS_BACKEND=dummy
